backend/requirements.txt
# spaCy + spaCy-transformers stack (kept as you chose Option 1)
spacy==3.7.4
spacy-transformers==1.3.5
transformers==4.36.2
# ↓ make these compatible with transformers 4.36.x
tokenizers==0.15.2
huggingface-hub>=0.19.3,<1.0

# (rest of your file)
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
redis==5.0.8
mlflow==2.16.0
python-dotenv==1.0.1
loguru==0.7.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
shap==0.46.0
captum==0.7.0
python-multipart==0.0.9
